+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMEdge-test.js
@@ -54,37 +54,6 @@ describe('ReactFlightDOMEdge', () => {
     use = React.use;
   });
 
-  function passThrough(stream) {
-    // Simulate more realistic network by splitting up and rejoining some chunks.
-    // This lets us test that we don't accidentally rely on particular bounds of the chunks.
-    return new ReadableStream({
-      async start(controller) {
-        const reader = stream.getReader();
-        let prevChunk = new Uint8Array(0);
-        function push() {
-          reader.read().then(({done, value}) => {
-            if (done) {
-              controller.enqueue(prevChunk);
-              controller.close();
-              return;
-            }
-            const chunk = new Uint8Array(prevChunk.length + value.length);
-            chunk.set(prevChunk, 0);
-            chunk.set(value, prevChunk.length);
-            if (chunk.length > 50) {
-              controller.enqueue(chunk.subarray(0, chunk.length - 50));
-              prevChunk = chunk.subarray(chunk.length - 50);
-            } else {
-              prevChunk = chunk;
-            }
-            push();
-          });
-        }
-        push();
-      },
-    });
-  }
-
   async function readResult(stream) {
     const reader = stream.getReader();
     let result = '';
@@ -143,69 +112,5 @@ describe('ReactFlightDOMEdge', () => {
     );
     const result = await readResult(ssrStream);
     expect(result).toEqual('<span>Client Component</span>');
-  });
-
-  it('should encode long string in a compact format', async () => {
-    const testString = '"\n\t'.repeat(500) + '🙃';
-    const testString2 = 'hello'.repeat(400);
-
-    const stream = ReactServerDOMServer.renderToReadableStream({
-      text: testString,
-      text2: testString2,
-    });
-    const [stream1, stream2] = passThrough(stream).tee();
-
-    const serializedContent = await readResult(stream1);
-    // The content should be compact an unescaped
-    expect(serializedContent.length).toBeLessThan(4000);
-    expect(serializedContent).not.toContain('\\n');
-    expect(serializedContent).not.toContain('\\t');
-    expect(serializedContent).not.toContain('\\"');
-    expect(serializedContent).toContain('\t');
-
-    const result = await ReactServerDOMClient.createFromReadableStream(
-      stream2,
-      {
-        ssrManifest: {
-          moduleMap: null,
-          moduleLoading: null,
-        },
-      },
-    );
-    // Should still match the result when parsed
-    expect(result.text).toBe(testString);
-    expect(result.text2).toBe(testString2);
-  });
-
-  // @gate enableBinaryFlight
-  it('should be able to serialize any kind of typed array', async () => {
-    const buffer = new Uint8Array([
-      123, 4, 10, 5, 100, 255, 244, 45, 56, 67, 43, 124, 67, 89, 100, 20,
-    ]).buffer;
-    const buffers = [
-      buffer,
-      new Int8Array(buffer, 1),
-      new Uint8Array(buffer, 2),
-      new Uint8ClampedArray(buffer, 2),
-      new Int16Array(buffer, 2),
-      new Uint16Array(buffer, 2),
-      new Int32Array(buffer, 4),
-      new Uint32Array(buffer, 4),
-      new Float32Array(buffer, 4),
-      new Float64Array(buffer, 0),
-      new BigInt64Array(buffer, 0),
-      new BigUint64Array(buffer, 0),
-      new DataView(buffer, 3),
-    ];
-    const stream = passThrough(
-      ReactServerDOMServer.renderToReadableStream(buffers),
-    );
-    const result = await ReactServerDOMClient.createFromReadableStream(stream, {
-      ssrManifest: {
-        moduleMap: null,
-        moduleLoading: null,
-      },
-    });
-    expect(result).toEqual(buffers);
   });
 });
