+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMForm-test.js
@@ -145,113 +145,4 @@ describe('ReactFlightDOMForm', () => {
     expect(result).toBe('hello');
     expect(foo).toBe('bar');
   });
-
-  // @gate enableFormActions
-  it('can submit an imported server action without hydrating it', async () => {
-    let foo = null;
-
-    const ServerModule = serverExports(function action(formData) {
-      foo = formData.get('foo');
-      return 'hi';
-    });
-    const serverAction = ReactServerDOMClient.createServerReference(
-      ServerModule.$$id,
-    );
-    function App() {
-      return (
-        <form action={serverAction}>
-          <input type="text" name="foo" defaultValue="bar" />
-        </form>
-      );
-    }
-
-    const ssrStream = await ReactDOMServer.renderToReadableStream(<App />);
-    await readIntoContainer(ssrStream);
-
-    const form = container.firstChild;
-
-    expect(foo).toBe(null);
-
-    const result = await submit(form);
-
-    expect(result).toBe('hi');
-
-    expect(foo).toBe('bar');
-  });
-
-  // @gate enableFormActions
-  it('can submit a complex closure server action without hydrating it', async () => {
-    let foo = null;
-
-    const serverAction = serverExports(function action(bound, formData) {
-      foo = formData.get('foo') + bound.complex;
-      return 'hello';
-    });
-    function App() {
-      return (
-        <form action={serverAction.bind(null, {complex: 'object'})}>
-          <input type="text" name="foo" defaultValue="bar" />
-        </form>
-      );
-    }
-    const rscStream = ReactServerDOMServer.renderToReadableStream(<App />);
-    const response = ReactServerDOMClient.createFromReadableStream(rscStream, {
-      ssrManifest: {
-        moduleMap: null,
-        moduleLoading: null,
-      },
-    });
-    const ssrStream = await ReactDOMServer.renderToReadableStream(response);
-    await readIntoContainer(ssrStream);
-
-    const form = container.firstChild;
-
-    expect(foo).toBe(null);
-
-    const result = await submit(form);
-
-    expect(result).toBe('hello');
-    expect(foo).toBe('barobject');
-  });
-
-  // @gate enableFormActions
-  it('can submit a multiple complex closure server action without hydrating it', async () => {
-    let foo = null;
-
-    const serverAction = serverExports(function action(bound, formData) {
-      foo = formData.get('foo') + bound.complex;
-      return 'hello' + bound.complex;
-    });
-    function App() {
-      return (
-        <form action={serverAction.bind(null, {complex: 'a'})}>
-          <input type="text" name="foo" defaultValue="bar" />
-          <button formAction={serverAction.bind(null, {complex: 'b'})} />
-          <button formAction={serverAction.bind(null, {complex: 'c'})} />
-          <input
-            type="submit"
-            formAction={serverAction.bind(null, {complex: 'd'})}
-          />
-        </form>
-      );
-    }
-    const rscStream = ReactServerDOMServer.renderToReadableStream(<App />);
-    const response = ReactServerDOMClient.createFromReadableStream(rscStream, {
-      ssrManifest: {
-        moduleMap: null,
-        moduleLoading: null,
-      },
-    });
-    const ssrStream = await ReactDOMServer.renderToReadableStream(response);
-    await readIntoContainer(ssrStream);
-
-    const form = container.firstChild;
-
-    expect(foo).toBe(null);
-
-    const result = await submit(form.getElementsByTagName('button')[1]);
-
-    expect(result).toBe('helloc');
-    expect(foo).toBe('barc');
-  });
 });
